Change SwiftApplication threadingModel from single to multi
I was having issues with single; and multi seems to be the default anyway.
The issue I had was 'Failed to initialize WindowsAppRuntimeInitializer:
Cannot change thread mode after it is set'. That code path gets called so
early in a WinUIBackend app's lifecycle, that there wasn't really anything
before that which could've been messing things up. Hopefully this change is
portable...